Add MMBF5103 JFET to my symbol library
So I've been using the J201 in my JFET input buffers, because I got that
JFET as apparently being a "traditional" JFET in guitar pedals.

But I also added the MMBF5103 to my component library since it seemed to
be a rather better JFET in many ways, with a reasonable V_{GSOFF} range
(-1.2V .. -2.7V) and I_{DSS} range (10mA .. 40mA).

In contrast, the V_{GSOFF} of the J201 is officially -0.3V .. -1.5V,
which is kind of tight for the simple "self-bias to ground" approach.

Now, I measured one of mine to be pretty much right in the middle at
-0.8V, and that matches what the spice file says, so at a guess with
modern manufacturing it's probably much tighter than what the data sheet
says, but whatever.

So the J201 actually works fine, but the MMBF5103 does seem to be a
"safer" choice for the simple self-biasing buffer setup, so let's use it
by default since I have them around.

The I_{DSS} range is very different (J201: 0.2mA .. 1mA) but I don't
think that matters in any of my designs. Again, I think the higher
I_{DSS} of the MMFB5103 potentially just makes it easier to use.

The MMBF5103 has a "high" gate reverse current at high temperatures that
looks relatively high, but that once more seems irrelevant for the kind
of stuff I do.

The two have the same footprint, and end upsimulating fairly similarly
with small signals and my kind of use, so I think they are - at least
for my current understanding and use cases - effectively a 1:1
replacement and I can play with either.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>