Update the Octaver circuit for MMBF5103 use
This not just switching the J201 into a MMBF5103, because when I did
that to verify that it's just a drop-in replacement, the simulation got
very unhappy indeed instead of saying "yup, same thing".

It looks like the extra current going through the 5103 ends up then
going through the subsequent BJT rectification stage, and the two stages
interact in bad ways causing oscillation on the phase-shifted signal.

That is easy enough to fix with just resistors between the stages.  But
I am also unhappy with the simulation voltages because the simulation
isn't seeing the capacitor on the base board, so I'll handle that
separately.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>