More octaver circuit impedance tweaking
I decided that the JFET phase splitter is the right way to go after all,
and that I just need to make sure the output impedance is low enough for
the following circuit.  Smaller transistors, more current, stiffer
signals.

But "more current" in this case is still in the "a couple of mA" range,
and this isn't battery-operated, so to quote Alfred E Neuman: "What, me
worry?"

I'm not particularly happy with that output side, and I'll probably
continue to play with this to get a better feel for the whole input and
output impedance of the different transistor circuits.  But maybe it's
time to make another test pcb just to see if this works in reality at
all.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>