Try to do a JFET compressor circuit.  Again.
No, really.  I don't seem to learn from past mistakes.  I keep thinking
that I can do a JFET compressor circuit, and they all fail miserably.

This one simulates in good ways for one particular amplitude-modulated
test signal.

Of course, one major problem with JFET's is that the physical
characteristics are all over the map, and this does *not* have any
tuning for that.  So reality will be very very different.

I doubt it works this time either, and I do need to have some way to
tune the JFETBIAS voltage.  But hey, it looks fine in simulation, and it
finally uses the rectification circuit for something that might not be
horrible noise.

Who am I kidding. It will _obviously_ just be horrible noise.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>