Add volume and level pots for compressor
The level pot might make the compressor able to deal with the reality of
JFET variability, at least within reason.

And obviously also in theory then allow people to actually set the level
at which things compress.

It's worth noting that the compression level is set purely based on the
input level. That's really stupid, since it really should be based on
feedback from the *output*, not the input. So high input levels might
then end up compressing to the point that the end result is less than
where medium input levels would have been.

But hey, if this things works even that well, that would be a small
miracle. The much more likely situation is that it will noisy, distort
horribly and not match my simulations at all.

My previous attempts at this do not instill confidence.

So I'll worry about smarter feedback when - and if - I ever get to that
bridge.

This also starts on the PCB layout. No components placed yet, just the
skeleton board with pots and connectors. But it's at the point where I
actually want to know if any of this works in reality, so...

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>