Add triangle wave for compressor testing
Because of how this odd "JFET as a voltage-controlled resistor in
parallel with a capacitor" circuit works, the whole ngspice AC signal
analysis doesn't work for this at all. Which limits me to just the
transient analysis.

So to do a poor-man's frequency sweep kind of thing, by using a 125Hz
triangle wave as an input and seeing how it distorts.

Does that show distortion? It sure does.

The thinking is that it's kind of a capacitor version of a resistive
T-network, with the reactance of the capacitors all frequency-dependent
in the same way (so scaling up and down together).

But then the JFET ends up being a non-frequency-dependent resistance
pulling the signal to ground, so now low frequencies are fundamentally
attenuated more than high frequencies (because low frequencies see more
reactance through the capacitors that aren't being bypassed with the
JFET).

And even *without* that JFET interaction, the capacitors obviously then
interact with the following resistors, so it all ends up acting like
highpass filter

Still, it acts better than any other circuit I've come up with. That's
not saying a lot, I know.

I think the frequency behavior can be improved on by making the caps
bigger - but I'm limited to 0.47uF as long as I want to use the C0G
MLCCs I have for the signal path.

But some distortion is inevitable, and this might just not be great for
bass guitar. So *IF* this circuit actually works, I suspect I'll just
have to play up the distortion as giving the sound "character".

I suspect the more likely scenario is that this will be just another
nasty-sounding thing.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>