Compressor: Use bigger capacitors in signal path
This just changes the 0.1uF caps in the signal path to 0.47uF ones.
That doesn't fix all ills, but it should pass lower frequency signals a
bit better.

This does make them 1206 footprints, and maybe the board will be a bit
tight.  I'm starting to look at the PCB next.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>