Compressor: do a first PCB layout
I thought this was going to be tighter than it was, so I started laying
out things unnecessarily densely.  Also, the lower left section (the
final signal amplification) would actually have been better off moved up
closer to the volume pot, but by the time I realized that I had tons of
space and shouldn't have squeezed things down there, it was mostly done
and I just didn't bother to change it to make a nicer layout.

I just want to know if this works at all, or if the simulation is just
being unreasonably optimistic.

If it actually works, I'll at a minimum want some kind of "it's now
starting to compress" LED or similar.  And I'd want to do it as a
feedback loop rather than as based purely on input volume (I did
actually simulate a first version of that and it looked doable, but
decided that I really want to get this one done and tested before taking
that step).

Put another way: if this is even remotely usable, I'll make bigger
changes that would result in a complete re-layout anyway, so no point in
worrying over-much about this not being a particularly nice layout.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>