Add my BAT54S as a symbol
Yeah, KiCad already has a BAT52S that looks about right, but it says
it's a SOT-323 in the name even though the footprint is actually the
correct one (SOT-23-3).  The two are not the same at all.  Don't you
love SMD name confusion?

Perhaps more importantly, this one has the simulation data associated
with it so that I don't need to look it up.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>