Compressor: re-do with feedback loop and a whole new approach
So as mentioned earlier I started looking into using a feedback loop
rather than base the compression purely on the input signal amplitude,
and as I was working on that, I realized that I had been thinking of the
JFET signal attenuation entirely the wrong way the whole time.

My confusion actually started way back with the old tremolo circuit that
I had so many problems with, and it all stems from just thinking that I
should use the JFET as a variable resistor to pull the signal down to
ground.

Yeah, yeah, that makes conceptual sense, but it's actually not the right
way to go about this all, because I realized that I was just making
things much harder than they needed to be.

It's much simpler and more practical to use a regular fixed resistor to
attenuate the signal by pulling it down to ground, and use the JFET to
actually _pass_ the signal.  After all, the default for the JFET is to
conduct, and R_{DS(ON)} is some reasonably small value like 50 Ohm.

So just turn things around, and it all becomes much more natural: with a
regular pull-down resistor in the kOhm range, the "normal" situation
(the JFET gate not actively pulled down) being that the JFET conducts
the signal and there is little to no attenuation.

Then when we want to attenuate the signal, we pull the gate down, and it
turns out that there's a very natural way to generate a negative signal
even without a dual power rail: just amplify the signal to have a large
enough voltage swing, and then do just a "voltage doubler" with two
capacitors and diodes.

It's just an inverting charge pump on the signal, in other words.  And
since it goes to the gate of the JFET, there's no need for any current.

Ta-daa, you have a nice negative voltage matching the signal amplitude
(with two diode drops, but just use a low-drop Schottky to minimize that
too).

And now the "add in half the input signal to the control signal" that I
had seen mentioned in JFET compressor discussions makes much more sense
too - it basically minimizes the issue of V_{GS} fluctuating with the
signal (and if I finally understood the JFET correctly, the "add half"
is simply because the source and drain are really symmetrical, so you
add the "half" of the signal that is associated with the source).

Anyway, this is all *much* more natural than my old circuit, and
simulates better too.

I may of course end up being disappointed, but I feel like it now
finally is coming together.

This does not lay out the PCB, but that should be a lot simpler too with
fewer components because I'm not doing silly things any more.

NOTE: various resistor and capacitor values are still mostly random from
various testing scenarios.  To make signal errors easier to see I was
lining up signals by tweaking hardcoded resistor values etc.  And while
I've spent _some_ time trying to get the attack and decay time constants
in the rough ballpark, they are mostly still random based on "this way I
can see the decay with my test signals" rather than anything necessarily
proper.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>