More compressor updates: add pots for attack / decay tuning etc
Lots of minor changes, but the user-visible one is that there is
(limited) control of the attack and delay now.  But the whole compressor
level setup has also changed, and the signal amplification is done a bit
differently too using the NPN/PNP pair in the NMB2227A.

Can you say "trying out random things"? I knew you could.

Let's get this manufactured and see if it actually works, since I won't
have time to mess with the design any more as the merge window opens and
my travel season starts.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>