Gaah - Fix stupid PNP direction mistake
I've done this *so* many times: I turn the PNP around by rotating it
like it's some random resistor, not my mirroring it around the Y-axis,
and I end up with the transistor the wrong way around without noticing
it.

And I did it *again*.

It didn't actually seem to matter for simulation, because this whole
"use a series pair" was just for the lulz and there wasn't any big
current or voltage gain expected.  And so the single NPN was perfectly
happy to drive it all.

But it's still very wrong indeed.

This just fixes the direction of the transistor, but I'll go look closer
if it had some unfortunate effect on the rest of my flawed results.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>