More compressor games
It turns ou tthat using the JFET to pass the signal (that I was so happy
about in commit c25f662: "Compressor: re-do with feedback loop and a
whole new approach") isn't so great after all.

In particular, it causes a kind of "zero-crossing distortion" where it
does compress the signal, but it does so by mainly compressing the
middle part, not the high ends.  Which looks all kinds of bad and
probably sounds worse.

I'm not 100% sure of the reason (I suspect there's some V_{DS} drop
effect), but clearly my clever trick was anything but.

Anyway, I do feel like I'm understanding how the JFET works much better
now, and I went back to the model with the JFET pulling the signal down,
and with my new-found understanding (famous last words) that looks a lot
better.

The whole "use a pot to set the voltage level" ended up making that
approach simpler too, because we don't now have to amplify the signal to
get that negative voltage with a diode/capacitor voltage inverter thing.
We can just adjust the level upwards to get a bigger V_{GS} drop..

This still does obviously amplify the signal, because we obviously don't
want to have a compressor that only makes the signal overall quieter.

I really would like to have a LED that shows when we start actually
compressing, but I'm still thinking about how to do that reasonably.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>