Update Octaver for bass frequencies
This is just updating the 0.1uF caps to 1uF, which gives much better
low-frequency behavior.  That does mean replacing the small 0704 form
factor with a 1206 one, so the pcb layout changes a bit.

It also means not using C0G caps, but let's be honest here: audiophile
this is not.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>