Update some compressor component values based on testing
I think I was being a bit too timid wrt the compressor, making it a bit
boring.  It definitely works, but it looks more interesting with a much
more aggressive compression ratio, and a more aggressive signal
amplification to match.

The level trim pot was also slow to respond because the capacitor to
keep it stable was fairly large.

Also, partially bypass the emitter resistor in the amplification stage
to give higher frequencies a bit more oomph - the AC signal analysis
claims they otherwise attenuated over 10kHz.

Most of this was tested with the old boards (just changed components).
And other parts are "this looks better in simulation" (eg the emitter
resistor bypassing).

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>