Add LS844 dual matched JFET symbol
I'm not loving the JFET level tuning in the compressor pedal, so I have
the glimmers of a plan.  Using a dual matched JFET I think I can use
self-biasing on one of the JFETs to get it into the interesting region,
and then use the resulting source voltage as the level voltage.

Will that work? Who knows, but this adds the symbol and the simulation
data for it so that I can try it out.

It sounds a bit wasteful to use one JFET just for that, but I've liked
the other dual transistor parts, so maybe I'll use this one for other
things too where I just want two JFETs even if I don't care about the
matching.

The LS844 seems like a perfectly cromulent JFET to use in general, with
a reasonable V_{GSOFF} range (-1...-3.5V) and I_{DSS} (1.5...15mA).

Compared to the MMBF5103, it has lower I_{DSS} and a slightly larger
V_{GSOFF} range, but if the matching helps with the variation the latter
should be fine.

If I read the datasheets right, the on-resistance of the LS844 seems to
be rather higher in the 100-200 Ohm range.  That too seems ok, but may
need tweaking when converting a circuit.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>