Update compressor for the LS844 dual JFET
This is just a quick conversion, not using the matching, just using the
LS844 as two independent JFETs.

Because I_{DSS} is lower, the source resistor on the input buffer is
bigger, so that if I_{DSS} is just 1.5mA the voltage drop over the
resistor can go up to 3.5V.

Similarly, tweak the limit resistor values to give a bigger limit
voltage range for the bigger V_{GSOFF} range of the LS844, and make the
negative feedback resistor divider bigger to account for higher R_{DSON}
of the LS844.

Also remove the partial bypassing of the amplification emitter resistor,
since the high-frequency damping of the plain base is no more.  Although
maybe it wants a brighter sound?

This seems to simulate ok, although the compression is certainly not
identical to the version with the MMBF5103.  That may be real, or may be
just different spice model default values and swamped by JFET
variations.

Anyway, I'll eventually try the "use self-biasing using the matching
JFETs to do automatic levels" eventually, but first I wanted to do the
minimal conversion and verify that at least the LS844 model seems to be
ok before doing any bigger surgery.

NOTE! This doesn't update the PCB layout for the changes.  Just doing
the schematic and simulation checks.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>