Compressor: base the compressor level voltage on matched JFET
This is an attempt at basing the compression level voltage on the
matched JFET, making the level knob hopefully more reliably sane in the
face of different JFET characteristics.

I'm not entirely convinced I shouldn't just use a trim pot, but I do
think this is a more interesting approach.  It basically picks the base
voltage from the self-biasing voltage of the matched JFET, guaranteeing
that the level voltage pot starts close to V_{GSOFF}.  The level pot is
then basically controlling the DC bias on top of that.

This also tries to make the actual attentuation circuit clearer and
hopefully better.  Both the JFET drain and source are biased to the
compression level voltage, while the base is biased to the signal peak
detector (the somewhat mis-named "RECT" net).

The source then also gets the signal, and when V_{GS} gets close to zero
(ie RECT gets close to LEVEL) the JFET starts conducting and the AC
signal is attenuated through the JFET to the capacitor to ground at the
drain.

The base gets half the AC signal (on top of the DC peak bias) as before
to minimize distortion (ie the AC signal at the base is halfway between
the AC signal at source and the lack of AC signal at drain).

This all seems to still simulate well, and the behavior of the level pot
seems to mostly make sense.  How well it works with actual random JFET
characteristics and whether the matching model works, I don't know.

I still need to actually lay out the PCB, but I think the schematic is
getting closer.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>