Update compressor PCB layout too
This does the new trial matches JFET thing as a board layout too.  I'm
really not convinced this was worth it, and the routing and layout are a
bit of a mess, but it's all part of the whole learnign experience, and I
do want to see if the self-biasing thing actually works in real life or
it just happens to look ok in simulation with the particular simulation
values.

.. and the usual noise from KiCad inexplicably changing randon things
when looking at files.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>