Add the beginnings of a dubious buffered base board
As part of looking at what I should try to use my MOSFETs for, I decided
I should finally look at a soft-switch buffered base board.

This is not it, but it starts with a very dubious push-on push-off
MOSFET circuit.  I doubt it really works, even if it simulates
beautifully.  But I don't simulate any debounce issues etc, although I
don't see why it wouldn't deal with that too.

All the real push-on push-off circuits I've seen are more complex than
this, which is the main reason why I doubt that my simple circuit really
works.  It's a bit *too* simple.  Everybody uses some flip-flop and/or
schmitt trigger or other fancier things.

But hey, maybe my google-fu is weak and I just didn't find the right
thing and there are even simpler ways to do this.  Plus I admittedly
came at this from an odd direction of "what can I do with discrete
MOSFETs".  Sane people don't do that.

Anyway, the basic of the push-on push-off circuit is - as always - a
capacitor that alternates between being discharged and charged up.  Then
two transistors that reinforce each other's state.

In this case, the capacitor is connected to the gate of the PMOSFET
through the momentary switch.  So if you press the switch and the
capacitor was discharged, the PMOSFET starts conducting, while if you
press the switch and the capacitor was charged, the PMOSFET turns off.

The rest of the circuit is just the complementary NMOSFET that then
follows the PMOSFET state and re-inforces whatever state it was in, and
the resistors to make it all work sanely.  The only important thing is
to make sure the resistors have sane values to not do bad things while
the switch is pressed.

It all simulates ok, and I don't see why it wouldn't work.

Anyway, this *only* does that silly push-on push-off logic.  It isn't
actually hooked up any way yet.  Not even the LED to show the state,
much less any JFETs to switch the actual signal around.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>