Wire the soft-switch up to actually switch signal and LED
Surprisingly, this actually seems to work fine.  In simulation only, of
course.

And not surprisingly, according to the same simulation, it has an
impressively horrible pop at switching time.  This very much reminds me
of the broken "Tremolo doubling as a metronome" thing, which was due to
the exact same thing: very harsh and sudden driving of the JFET gate.

It's made worse by the MOSFET switching having the same thing.  Big
voltage spike when turning off, presumably due to capacitive parasitics
or whatever.

Possibly helped by adding a resistor at the gate or something, but
unlike the tremolo, the switching event is rare.  So right now I simply
don't care.  I'm just _simulating_ switching things at a 5Hz rate, so it
shows up like a sore thumb.

I haven't done layout, and I haven't added a real switch yet, so this is
all just simulation.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>