Buffered base: add RF filter after JFET switches
It turns out that getting rid of the switching pop seems to be as simple
as that.

Maybe time to do PCB layout and see if reality matches simulation?

I do think I should do the buffer that this doesn't do despite the name.
That was the intent, but then it worked so well without it..  An op-amp
would be the sane thing to do, but since I've been on a "discrete
components" tear, I suspect I'd just do a JFET voltage follower.

But then I feel like I should make up for the inevitable signal loss.
Or use a common source thing that does a phase reversal? That seems
wrong too.  Perhaps do another dual JFET for a common source both at
input and output?

Decisions, decisions.

Anyway, this doesn't do any of that.  This just gets rid of the
switching pop in simulation.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>