Split the buffered base over multiple schematic pages, do PCB layout
So this thing is still mis-named, in that it doesn't actually buffer the
signal at all, it only does the transistor soft-switching.  But I _will_
add buffering at some point, it's just not the interesting part of this
all, since I've done that part before.

But because I wrote explanations for the push-on / push-off circuit, and
because I split things up into small parts, the schematic was getting
pretty crowded.  So I decided to split up the actually interesting part
of the schematic - the switching side - onto a schematic page of its
own.

And while doing that, I moved the simulation infrastructure to its own
page too.

That leaves the toplevel page pretty empty with mainly just the
connectors and the basic signal and power filtering, but that's where
the buffers will go eventually.

And because it simulates fine, I want to actually see if it works in
reality.  So this also does the PCB layout.  There's tons of room left
for any future buffering circuit, whether I then do it with discrete
JFETs or decide that op-amps are the sane thing here.

I used a random regular tactile switch - which I don't actually have in
my parts library.  I'll just wire up two wires and test it that way (and
I do have some momentary SPDT stomp switches around that I can attach to
the wires).

But longer-term, one of the points of this kind of electronic switching
is that you can do it multiple ways, and aren't limited to a physical
switch - or indeed just a single one.  So maybe I'll do some remote
switching thing, or play around with using a hall effect sensor and a
magnet, or whatever.

There are always new things to learn about...  But let's make sure the
basics of this thing actually works first.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>