Add LT3462A symbol
I'm going to experiment with a negative power rail one more time, and in
the LT3462A I found a DC/DC converter that looks a lot more pleasant
than the ones I've tried before.

Not only is it a nice TSOT-23-6 footprint, but it has an integrated
schottky diode so it's easy to use.  And it's a fixed frequency dual
inductor inverter design, so it should be reasonably quiet in the audio
frequency.

I'm not planning to bother with LDOs to clean up the power rails,
because the reason for the power rail would be the op-amp for buffering,
and the PSRR of any quality opamp should make it all irrelevant.

I dunno.  Every time I've tried this, it's just been noise hell, so this
is probably a bad idea.  Even when you can't actually _hear_ the noise,
it ends up just messing with the signal on the scope and making it
harder to judge whether the circuit is doing well or not.  But I keep
trying.

What was the definition of insanity again? I forget.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>