Buffered base: use negative rail for JFET switching
This _still_ doesn't do any actual buffering, but with the negative
rail, we now no longer need to rebase the signals to some mid-point
voltage, and we can just leave the signals ground-references and use the
negative rail for switching.

That streamlines the signal path, and will make it easier to just insert
an opamp or two in there too.  But I remain more interested in getting
the switching side sorted out properly, so that's the primary goal here.

This tweaks the push-on push-off circuit a bit too, since taht one now
needs to switch between positive and negative voltages rather than VIN
and ground.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>