Update the buffered base board
The JFET switching worked fine, but since this all started as an
exercise in using MOSFETs, let's just go all in on that, and use the
MOSFETs for switching the signal too.

Due to the body diode, you need two MOSFETs in series for the signal
switching, so it's four MOSFETs instead of two JFETs, but despite that
the end result is actually simpler.  And those dual MOSFET things are
common and dirt cheap, unlike the special matched JFET package.

Since I can just continue to use the dual complementary MOSFETs, I can
use PMOS for one signal and NMOS for the other, and share all the gate
circuit.  Plus no need for a diode to protect the gate current messing
with the signal.

Also update the resistor value in the switching circuit and make the
time constant be 100ms.  That seems to make it more reliable in testing,
although I'm not convinced my testing with the original value was all
that great.  The soft-switch circuit does seem to work, though.

Also, since the LED pin on the board connectors has never been used,
let's just re-purpose that as a negative voltage pin for the effect
board.  If some effect board wants to touch the LED status, it might as
well just connect to the LED itself.  Having access to a dual voltage
rail would seem to be a lot more important.

Finally - and probably most importantly for actual sound and current
effects boards - add a low-pass filter on the voltage lines.  Because as
expected, the negative power supply did add a lot of noise to the VIN
line.

The base board itself doesn't actually care, but the effects boards do.

An actual LDO would be better, but I don't have any simple fixed ones in
my parts library (except for the 5V one, which seems a bit sad).
Current effects boards don't use a lot of current and only care about
the voltage being stable, so let's try this simple model.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>