Create a pedal design based on the Electrosmith Daisy Seed platform
This tries to make the pedal actually useful by making the audio side
use the Daisy Seed from Electrosmith [1]

I've had the board for the last six months, but up until now didn't feel
ready to use it.

I even built the Hothouse pedal kit around it, and exchanged some emails
with both Andrew Ikenberry (Electrosmith) and Ricky Sheaves (of Hothouse
pedal from Cleveland Music Co), but never even flashed it up to test.

But now that I'm getting bored with trying variations of the basic
analog base board, let's switch things up a bit and do a base board for
the Daisy Seed instead.

Because assuming I don't mess anything up, it's obviously going to sound
much better and be a lot more flexible than some silly analog distortion
circuit from half a century ago.

[ This is just a worse version of the Hothouse pedal, using a crazy dual
  JFET thing for input buffering just because the sane thing that the
  Hothouse does - using a much cheaper and more capable op-amp - feels a
  bit boring.

  So unlike the Hothouse, this is a mono board and objectively less
  useful. Go get the real thing from [2]. This is just another board in
  the endless series called "Linus is trying new ways to do pointless
  things" ]

So this takes my traditional base board, turns it around so that the
power and audio connectors go on the bottom of the board, and puts the
control pots on the top side.

That way all the mechanical hardware is on the base board.

I think this is what I probably should have done originally, but it
forces a fixed layout for pots and switches, which didn't appeal to me
in the original design, because the pot connections depend on the
effect.

However, since the enclosures are such a pain, a fixed layout is
actually probably a good thing, and with the Electrosmith Daisy Seed,
the pots are all connected the same way - as resistor diviers to ADC
inputs.

And since this puts all the mechanicals on that base, the effect board
now has a lot more space.

Which wasn't an issue with the original trivial effects, but now that
I'm trying to fit the whole Daisy Seed on that thing, it suddenly
becomes a major issue.

Will it fit? Naah - the Daisy Seed board is just too high in the Z
dimension. But I'm thinking about it. There's a distant plan.

This also updates the 1590A footprint for the new fixed 3-pot layout and
for the new board connector placement. And incidentally also fixes the
location for the LED and the switch that had been moved in a long time
ago.

The effects board then pretty much just connects the pin headers from
the connector board to the Daisy Seed. There's a capacitor in there too,
but that's literally it.

This does it all with a custom Daisy Seed symbol and footprint.

There were already versions of that in the KiCad libraries, but I'm not
a fan. The standard KiCad version has the whole Daisy seed as a keep-out
area, and I want to be able to put low-profile things under the board
without DRC complaints.

My custom one also ends up just using the Hothouse pin names (so "POT1"
for the first potentiometer, rather than GPIO16/ADC1). So it's a bit
clearer what is going on by being specific to my particular use.

Link: https://electro-smith.com/ [1]
Link: https://clevelandmusicco.com/hothouse-diy-digital-signal-processing-platform-kit/ [2]
Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>