Add input signal re-biasing to the Daisy Seed after all
Yes, the Daisy Seed ends up AC-coupling the input, and will re-bias it
to the mid-point of the analog voltage range for the audio codec.

But thinking some more about it, I started worrying about the capacitor
used. In particular, it might very well be a low-voltage one, and the
dual JFET input will have a buffered output voltage in the 5V+ range.

Simulation puts the signal level centered at 7.4V, but who knows what an
actual LS844 will be.  For all I know the dual JFET thing really would
want more of a voltage range than a 9V positive supply and it will not
work very well in real life, but we'll see.

[ I can always go back to the single MMBF5103 as an input buffer, that
  worked fine and is more than sufficient for a guitar pedal -
  particularly one of my design.

  The whole "let's used two matched JFET's" is really me being rather
  silly, after having seen reports of it having particularly good THD by
  Rod Elliott (ESP) [1] and Erno Borbely [2].

  ESP uses the matched JFET as a current source, Erno Borbely as a
  current sink. Mostly it shouldn't make any difference, but I randomly
  did the current sink thing. And that does end up with the signal at a
  higher voltage ]

Regardless, it's entirely possible that the AC coupling on the Daisy
Seed uses a capacitor that is rated for something like 6.3V, which is
common for smaller SMT caps.  That would make sense, since the
assumption is an incoming signal around ground, with the center of the
analog voltage at 1.6V.

So my assumption that the AC coupling of the Daisy Seed would make it
unnecessary to do any AC coupling on the base board signal was clearly
horribly flawed.  Let's fix it.

I wonder what the next horribly flawed idea is?

Link: https://sound-au.com/articles/jfet-design.htm#s5 [1]
Link: https://pearl-hifi.com/06_Lit_Archive/14_Books_Tech_Papers/Borbely_Erno/JFETS_The_New_Frontier.pdf [2]
Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>