Try attaching the Daisy Seed directly on the base board
Ok, so to fit this thing, I needed to play some games, and also get rid
of the third potentiometer.

So now it only has two pots, and the Daisy Seed needs to be the
unsoldered kind, where you solder low-profile pin headers to the "wrong"
side of the board.

IOW, the Daisy Seed gets mounted upside down on the top of the base
board, and then it kind of fits.

I say "kind of", because this is all without a stomp switch.  I haven't
decided on what I want to do about that.  The switch connector is just
holes for a pin header - or more likely, two wires.

I do think I should just give up on the 1950A form factor, which would
make all these games a complete non-issue.  But it's not like the whole
thing makes any sense to begin with, so why start being sensible now?

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>