Add very incomplete and broken stm32 board
The potential plan is to maybe try to actually recreate a subset of what
the Daisy Seed does, and have it populated at jlcpcb.

Because I'm not excited about hand-soldering the kinds of components
this will need (although so far I've picked stuff that should be doable,
so maybe I change my mind and just go for a stencil).

This is probably all kinds of broken garbage.  And it's very incomplete.
I've done a MCU board before, but that was years ago.  I've never done
one of the stm32 ones.

Not yet done: SWD, proper USB, and the PCM3060.  And whatever else I'm
missing and haven't even thought of yet.

Also, since I've picked a stm32 part that I could maybe solder at home,
it's one that doesn't have enough pins brought out for SDRAM.  So I'd be
limited to the internal 1MB of RAM.

Put another way: don't for a moment think that this replaces a real
Daisy Seed.  That board works, was designed by somebody who knew what
they were doing, and is clearly superior.

In contrast, this is me doing silly things again, in the name of trying
to learn something new.

In other and more uplifting news: the dual-JFET LS844 input circuit on
the base board actually does work, and seems to work very well indeed.

I've only built one board, but that one looked very linear and clean on
a triangle wave all the way between 10mVp-p to 4Vp-p in the whole
20Hz-20kHz band.

In fact I checked it from 10Hz all the way up to 1Mhz.

And yes, it was recognizably a signal even at 1MHz.  Admittedly at that
point it looked much more like a sine-wave and had attenuated quite a
lot.  There's an intentional RF filter in the input path, so that's
basically expected.

In fact that signal looked *so* good that I'm suspicious that I did
something wrong.  But at 4Vp-p I followed the input and the output on
the oscillosope all the way from 20Hz to 20kHz, and they looked visually
indistinguishable and the signals were on top of each other from 70Hz up
(20-70Hz you'd see the triangle wave bowing a bit due to AC coupling).

At 10mVp-p, you could definitely see noise, but I'm not sure how much of
it comes from the actual circuits, and how much of it is from cables (ie
comparing a nice direct BNC cable from the signal generator to the
oscilloscope with one that goes through two guitar patch cables and the
1/4" jacks: not exactly equivalent signal integrity).

The HF noise is maybe 2mVp-p (looks like ~1mV if I limit the scope
bandwidth to 20MHz), so it's a noticeable fraction of the 10mV signal,
but on the whole this still looks like my best input path by far.

The 10mVp-p signal still looks good at 100kHz (and also still
recognizably a signal at 1MHz).

NOTE! That's just the input signal path when jumpered directly to the
output.  When adding the Daisy Seed on the test board into the path it
is a different thing entirely, of course.  *That* noise looks more like
100mVp-p, and I have to look at a 1Vp-p signal to make it really
recognizable.

The Daisy Seed still *sounds* pretty good, because the noise seems to be
mainly HF noise at 450kHz or something - I suspect it ends up being the
3.3V switching power supply on the Daisy Seed that presumably changes
frequency based on load.

It just looks horrible on the oscilloscope, but sounds way better than
any of my analog effects did.

Also, looking at the waveform, the Daisy Seed delay - without any effect
active - is about 5.2ms.  That's with the Hothouse "ShimmerVerb" effect
loaded, it might depend on some setup.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>