Add aggressive low-pass filter to the Daisy Seed base board output
This is from me looking at the HF noise from the Daisy Seed on the
oscilloscope, and then checking what the Hothouse padal looks like.

The Hothouse actually looks really clean - kudos to Ricky Sheaves.  The
reason seems to be a rather aggressive low-pass filter with the 3dB
point at 22kHz.  That will attenuate some of the audio band too, but
looks like it's the right thing to do considering the noise.

(Of course, the Hothouse also does an op-amp to buffer the end result.
My silly discrete component board does not.  That will affect the end
result too, but the output impedance should already be pretty low for a
guitar pedal even without the extra output buffer).

Change the sim to have a simulated 100 Ohm impedance signal output from
the Daisy Seed so that the AC simulation makes sense.  Without that this
just resulted in nonsensical output filtering.

While looking at the AC response, also change the input caps to 4.7uF -
I think that's what I used for the test board, and you definitely want
something bigger than 0.47uF for bass response.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>